+++ test/package.test.js
@@ -21,11 +21,7 @@ describe('Usage `ansis` npm package', () => {
       '\x1b[38;2;255;117;209mextend: ansis2.pink\x1b[39m\n' +
       '\x1b[3m\x1b[38;2;255;171;64mextend: ansis2.italic.orange\x1b[39m\x1b[23m\n' +
       'colored:  \x1b[32mgreen text\x1b[39m\n' +
-      'striped:  green text\n' +
-      '--- OLD  named import: \x1b[33mconst { magenta, cyan, blue } = require(\'ansis/colors\');\x1b[39m\n' +
-      '\x1b[35mmagenta\x1b[39m\n' +
-      '\x1b[36m\x1b[1mcyan bold\x1b[22m\x1b[39m\n' +
-      '\x1b[34m\x1b[3mblue italic\x1b[23m\x1b[39m';
+      'striped:  green text';
     expect(esc(received)).toEqual(esc(expected));
   });
 
@@ -45,11 +41,7 @@ describe('Usage `ansis` npm package', () => {
       '\x1b[38;2;255;117;209mextend: ansis2.pink\x1b[39m\n' +
       '\x1b[3m\x1b[38;2;255;171;64mextend: ansis2.italic.orange\x1b[39m\x1b[23m\n' +
       'colored:  \x1b[32mgreen text\x1b[39m\n' +
-      'striped:  green text\n' +
-      '--- OLD  named import: \x1b[33mimport { magenta, cyan, blue } from \'ansis/colors\';\x1b[39m\n' +
-      '\x1b[35mmagenta\x1b[39m\n' +
-      '\x1b[36m\x1b[1mcyan bold\x1b[22m\x1b[39m\n' +
-      '\x1b[34m\x1b[3mblue italic\x1b[23m\x1b[39m';
+      'striped:  green text';
     expect(esc(received)).toEqual(esc(expected));
   });
 });